+++ cargo-concordium/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Non-existing directories in paths provided to the following arguments for when running `cargo concordium build` will now be created instead of causing an error: `--out`, `--schema-out`, `--schema-json-out`, `--schema-base64-out`.
   Likewise for the `--out-bin` and `--out-json` arguments provided to `cargo concordium run init` and `cargo concordium run update`.
+- Fix a bug where `cargo concordium` was unable to determine the smart contract package if the package was part of a Cargo workspace.
 
 ## 2.7.1